Armor penetration
This commit brings the penetration solver, which works pretty good and replicates 1:1 behavior with War Thunder

The depth solver for source props is currently in expensive mode for the most accurate depth estimations of ray-prop interactions

Ricochet is not an emergent behavior, so it will come later. However, angled armor does produce correct effective thickness results.

The solver is hardcoded for the M61 shell, but soon it'll be replaced by the ammunition system.

It's also hardcoded at everything being dense RHA, which will need some changes.